Social_Media_Project/requirements.txt
amqp==5.1.1
annotated-types==0.6.0
anyio==4.2.0
asgiref==3.7.2
async-timeout==4.0.3
backports.zoneinfo==0.2.1
billiard==4.1.0
celery==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.0
Django==4.2.4
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-filter==24.3
django-timezone-field==6.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
exceptiongroup==1.2.0
fastapi==0.109.2
gunicorn==21.2.0
h11==0.14.0
idna==3.4
kombu==5.3.1
numpy==1.24.4
packaging==23.2
pandas==2.0.3
peppercorn==0.6
Pillow==10.0.1
prompt-toolkit==3.0.39
pydantic==2.6.1
pydantic_core==2.16.2
PyJWT==2.8.0
python-crontab==3.0.0
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3
redis==5.0.0
requests==2.31.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
starlette==0.36.3
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.0.4
uvicorn==0.27.1
vine==5.0.0
wcwidth==0.2.6
